initial rework of the original acceptance tests
